requirements.txt
flask>=3.0.0
openai>=1.30.0
python-dotenv>=1.0.1
opencv-python>=4.9.0
numpy>=1.26.0
